pyproject.toml
[project]
name = "vehiclepass"
version = "0.1.0"
description = "Experimental implementation of the FordPass API."
readme = "README.md"
authors = [
    { name = "Austin de Coup-Crank", email = "austindcc@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pint>=0.24.4",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ruff>=0.11.2",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "RUF", # ruff-specific rules
    "W",   # pycodestyle warnings
    "D",   # pydocstyle
]

exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint.isort]
known-first-party = ["vehiclepass"]
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"
